projet-backend-django-minor-main/requirements.txt
django 
djangorestframework
channels==3.0.4
channels-redis
celery[redis]
django-celery-beat
django_celery_results
Random-Word
pyyaml
requests
pillow
setuptools
django-cors-headers
python-dotenv
gunicorn
whitenoise
dj-database-url
django-filter
djangorestframework-simplejwt
psycopg2
scipy